fix: setup for both rncli & rnef
