+++ packages/module-federation-metro/react-native.config.js
@@ -1,3 +1,14 @@
-const commands = require("./commands");
+const {
+  bundleFederatedRemote,
+  bundleFederatedRemoteOptions,
+} = require("./commands");
 
-module.exports = { commands };
+const bundleMFRemoteCommand = {
+  name: "bundle-mf-remote",
+  description:
+    "Bundles a Module Federation remote, including its container entry and all exposed modules for consumption by host applications",
+  func: bundleFederatedRemote,
+  options: bundleFederatedRemoteOptions,
+};
+
+module.exports = { commands: [bundleMFRemoteCommand] };
