+++ packages/module-federation-metro/src/commands/utils/loadMetroConfig.ts
@@ -1,15 +1,15 @@
 import path from "node:path";
 import type { ConfigT, InputConfigT, YargArguments } from "metro-config";
 import { loadConfig, mergeConfig, resolveConfig } from "metro-config";
-import { BundleRemoteConfig } from "../types";
+import { BundleFederatedRemoteConfig } from "../types";
 import { CLIError } from "./cliError";
 
 function getOverrideConfig(
-  ctx: BundleRemoteConfig,
+  cfg: BundleFederatedRemoteConfig,
   config: ConfigT
 ): InputConfigT {
   const resolver: Partial<ConfigT["resolver"]> = {
-    platforms: [...Object.keys(ctx.platforms), "native"],
+    platforms: [...Object.keys(cfg.platforms), "native"],
   };
 
   return {
@@ -19,8 +19,8 @@ function getOverrideConfig(
         ...(config.serializer?.getModulesRunBeforeMainModule?.(entryFilePath) ||
           []),
         require.resolve(
-          path.join(ctx.reactNativePath, "Libraries/Core/InitializeCore"),
-          { paths: [ctx.root] }
+          path.join(cfg.reactNativePath, "Libraries/Core/InitializeCore"),
+          { paths: [cfg.root] }
         ),
       ],
     },
@@ -28,10 +28,10 @@ function getOverrideConfig(
 }
 
 export default async function loadMetroConfig(
-  ctx: BundleRemoteConfig,
+  cfg: BundleFederatedRemoteConfig,
   options: YargArguments = {}
 ): Promise<ConfigT> {
-  const cwd = ctx.root;
+  const cwd = cfg.root;
   const projectConfig = await resolveConfig(options.config, cwd);
 
   if (projectConfig.isEmpty) {
@@ -40,7 +40,7 @@ export default async function loadMetroConfig(
 
   const config = await loadConfig({ cwd, ...options });
 
-  const overrideConfig = getOverrideConfig(ctx, config);
+  const overrideConfig = getOverrideConfig(cfg, config);
 
   return mergeConfig(config, overrideConfig);
 }
